+++ AmplifyPlugins/API/AWSAPICategoryPluginTests/AWSAPICategoryPluginTestBase.swift
@@ -31,7 +31,10 @@ class AWSAPICategoryPluginTestBase: XCTestCase {
 
     override func setUp() {
         apiPlugin = AWSAPIPlugin()
-        authService = MockAWSAuthService()
+
+        let authService = MockAWSAuthService()
+        self.authService = authService
+
         do {
             let endpointConfig = [apiName: try AWSAPICategoryPluginConfiguration.EndpointConfig(
                 name: apiName,
@@ -39,10 +42,17 @@ class AWSAPICategoryPluginTestBase: XCTestCase {
                 region: region,
                 authorizationType: AWSAuthorizationType.none,
                 authorizationConfiguration: AWSAuthorizationConfiguration.none,
-                endpointType: .graphQL)]
-            pluginConfig = AWSAPICategoryPluginConfiguration(endpoints: endpointConfig)
-            apiPlugin.configure(authService: authService,
-                                pluginConfig: pluginConfig)
+                endpointType: .graphQL,
+                apiAuthProviderFactory: APIAuthProviderFactory())]
+            let pluginConfig = AWSAPICategoryPluginConfiguration(endpoints: endpointConfig)
+            self.pluginConfig = pluginConfig
+
+            let dependencies = AWSAPIPlugin.ConfigurationDependencies(
+                pluginConfig: pluginConfig,
+                authService: authService,
+                subscriptionConnectionFactory: AWSSubscriptionConnectionFactory()
+            )
+            apiPlugin.configure(using: dependencies)
         } catch {
             XCTFail("Failed to create endpoint config")
         }