+++ AmplifyPlugins/DataStore/AWSDataStoreCategoryPlugin/Sync/RemoteSyncEngine+IncomingEventReconciliationQueueEvent.swift
@@ -34,6 +34,8 @@ extension RemoteSyncEngine {
     func onReceive(receiveValue: IncomingEventReconciliationQueueEvent) {
         switch receiveValue {
         case .initialized:
+            let payload = HubPayload(eventName: HubPayload.EventName.DataStore.subscriptionsEstablished)
+            Amplify.Hub.dispatch(to: .dataStore, payload: payload)
             remoteSyncTopicPublisher.send(.subscriptionsInitialized)
             stateMachine.notify(action: .initializedSubscriptions)
         case .started:
@@ -44,8 +46,10 @@ extension RemoteSyncEngine {
             }
         case .paused:
             remoteSyncTopicPublisher.send(.subscriptionsPaused)
-        case .mutationEvent(let mutationEvent):
+        case .mutationEventApplied(let mutationEvent):
             remoteSyncTopicPublisher.send(.mutationEvent(mutationEvent))
+        case .mutationEventDropped:
+            break
         }
     }
 }